+++ src/Utils/DataStructures/IndicesRangeSolidVector.h
@@ -4,28 +4,28 @@
 
 struct IndicesRangeSolidVector
 {
-	void EraseIndex(size_t index)
+	void EraseIndex(uint32_t index)
 	{
 		m_freeIndices.push_back(index);
 	}
 
-	size_t AllocateIndex()
+	uint32_t AllocateIndex()
 	{
 		if (m_freeIndices.empty())
 		{
 			ASSERT(m_maxIndexCreated + 1 > m_maxIndexLimit, "Indeces total count exceeded limit");
 			return ++m_maxIndexCreated;
 		}
-		size_t result = m_freeIndices.back();
+		uint32_t result = m_freeIndices.back();
 		m_freeIndices.pop_back();
 		return result;
 	}
 
-	std::vector<size_t> AllocateIndicesRange(size_t totalNum)
+	std::vector<uint32_t> AllocateIndicesRange(uint32_t totalNum)
 	{
-		std::vector<size_t> result;
+		std::vector<uint32_t> result;
 		result.reserve(totalNum);
-		for (size_t i = 0; i < totalNum; i++)
+		for (uint32_t i = 0; i < totalNum; i++)
 		{
 			if (m_freeIndices.empty())
 			{
@@ -34,7 +34,7 @@ struct IndicesRangeSolidVector
 			}
 			else
 			{
-				size_t index = m_freeIndices.back();
+				uint32_t index = m_freeIndices.back();
 				m_freeIndices.pop_back();
 				result.push_back(index);
 			}
@@ -42,30 +42,30 @@ struct IndicesRangeSolidVector
 		return result;
 	}
 
-	size_t EraseIndicesRange(size_t rangeBeginIndex, size_t totalNum)
+	uint32_t EraseIndicesRange(uint32_t rangeBeginIndex, uint32_t totalNum)
 	{
-		for (size_t i = rangeBeginIndex; i < rangeBeginIndex + totalNum; i++)
+		for (uint32_t i = rangeBeginIndex; i < rangeBeginIndex + totalNum; i++)
 		{
 			m_freeIndices.push_back(i);
 		}
 	}
 
-	size_t GetTotalSize()
+	uint32_t GetTotalSize()
 	{
 		return m_maxIndexCreated + 1;
 	}
 
-	size_t GetCurrentlyAllocatedIndicesNumber()
+	uint32_t GetCurrentlyAllocatedIndicesNumber()
 	{
 		return GetTotalSize() - m_freeIndices.size();
 	}
 
-	void SetMaxIndexLimit(size_t value)
+	void SetMaxIndexLimit(uint32_t value)
 	{
 		m_maxIndexLimit = value;
 	}
 
-	std::vector<size_t> m_freeIndices{};
-	size_t m_maxIndexCreated = -1;
-	size_t m_maxIndexLimit = ULLONG_MAX;
+	std::vector<uint32_t> m_freeIndices{};
+	uint32_t m_maxIndexCreated = -1;
+	uint32_t m_maxIndexLimit = ULLONG_MAX;
 };
